Added new projects
UI for launching automated tests on  different control types to practice
using the UI automation framework (with placeholder controls to add
server browsing/db creation and test case input later).
Also a project containing the UI that will be automated.
Practiced using IEnumerable to populate list box and started to try and
apply the UI framework to find the form.